minor changes everywhere, finished textdisplay

+++ subject.h
@@ -10,7 +10,7 @@ class Subject {
   void attach( Observer* o );
   void detach( Observer* o );
   void notifyObservers();
-  virtual char getState( int row, int col ) const = 0;
+  virtual char getState( size_t row, size_t col ) const = 0;
   virtual ~Subject();
 };
